1733 - Moved column selector to clean workplace utils into a class.
